requirements.txt
streamlit==1.28.2
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
pdf2image